+++ src/main/resources/schema.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS watch_catalogue (
     id VARCHAR(255) NOT NULL,
     name VARCHAR(255) NOT NULL,
-    unit_price INTEGER NOT NULL,
+    unit_price DECIMAL(10, 2) NOT NULL,
     discount_expression VARCHAR(255),
     PRIMARY KEY (id)
 );
